qcacld-3.0: Move rso related process to connection manager [PART 7]
Add new code to implement below functions for connection manager
roam part:
Filling below WMI cmd parameters related process:
WMI_ROAM_CONFIGURE_MAWC_CMDID
WMI_ROAM_BSS_LOAD_CONFIG_CMDID

Rename file wlan_cm_roam_public_srtuct.h name to
wlan_cm_roam_public_struct.h

CRs-Fixed: 2758130

+++ mlme/dispatcher/inc/wlan_mlme_public_struct.h
@@ -1577,8 +1577,8 @@ struct wlan_mlme_lfr_cfg {
 	bool wes_mode_enabled;
 	uint32_t mawc_roam_traffic_threshold;
 	uint32_t mawc_roam_ap_rssi_threshold;
-	uint32_t mawc_roam_rssi_high_adjust;
-	uint32_t mawc_roam_rssi_low_adjust;
+	uint8_t mawc_roam_rssi_high_adjust;
+	uint8_t mawc_roam_rssi_low_adjust;
 	uint32_t roam_rssi_abs_threshold;
 	uint8_t rssi_threshold_offset_5g;
 	uint8_t early_stop_scan_min_threshold;
